[9.0][FIX] stock_deposit: Adapt to OCA conventions.
