+++ stock_deposit/models/__init__.py
@@ -2,5 +2,8 @@
 # © 2016 Sergio Teruel <sergio.teruel@tecnativa.com>
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-from . import stock
+from . import stock_location
+from . import stock_warehouse
+from . import stock_picking
+from . import stock_quant
 from . import product
