+++ stock_deposit/views/stock_dashboard.xml
@@ -4,7 +4,7 @@
 <odoo>
 
 <record id="deposit_stock_picking_type_kanban" model="ir.ui.view">
-    <field name="name">depoasit.stock.picking.type.kanban</field>
+    <field name="name">deposit.stock.picking.type.kanban</field>
     <field name="model">stock.picking.type</field>
     <field name="inherit_id" ref="stock.stock_picking_type_kanban"/>
     <field name="arch" type="xml">
