+++ stock_deposit/views/stock_quant_view.xml
@@ -1,0 +1,53 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!-- © 2016 Sergio Teruel <sergio.teruel@tecnativa.com>
+     License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl-3). -->
+<odoo>
+
+<record id="deposit_quant_search_view" model="ir.ui.view">
+    <field name="name">deposit.stock.quant.search</field>
+    <field name="model">stock.quant</field>
+    <field name="inherit_id" ref="stock.quant_search_view"/>
+    <field name="arch" type="xml">
+        <xpath expr="//filter[@name='internal_loc']" position="after">
+            <!-- we need and operator -->
+            <separator/>
+            <filter name='deposit_loc'
+                    string="Deposit Locations"
+                    domain="[('location_id.deposit_location', '=', True)]"/>
+            <filter name='deposit_loc_exclude'
+                    string="Deposit Locations Excluded"
+                    domain="[('location_id.deposit_location', '!=', True)]"/>
+        </xpath>
+    </field>
+</record>
+
+<record id="deposit_product_open_quants" model="ir.actions.act_window">
+    <field name="name">Stock In Deposit</field>
+    <field name="context">{'search_default_internal_loc': 1,
+                           'search_default_locationgroup':1,
+                           'search_default_deposit_loc': 1}
+    </field>
+    <field name="domain">[('product_id', '=', active_id)]</field>
+    <field name="res_model">stock.quant</field>
+</record>
+
+<record id="deposit_quants_act" model="ir.actions.act_window">
+    <field name="name">Deposited Quants</field>
+    <field name="res_model">stock.quant</field>
+    <field name="view_type">form</field>
+    <field name="view_mode">tree,form</field>
+    <field name="context">{'search_default_internal_loc': 1,
+                           'search_default_deposit_loc': 1,
+                           'group_by': ['product_id', 'location_id']}
+    </field>
+    <field name="help">
+        This analysis gives you a fast overview on the current stock level of
+        your products in deposit locations.
+    </field>
+</record>
+
+<menuitem id="menu_deposit_quants" name="Deposited Inventory Valuation"
+          parent="stock.menu_warehouse_report" sequence="125"
+          action="deposit_quants_act"/>
+
+</odoo>
