+++ stock_deposit/views/stock_view.xml
@@ -3,45 +3,6 @@
      License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl-3). -->
 <odoo>
 
-<record id="deposit_view_location_form" model="ir.ui.view">
-    <field name="name">deposit.stock.location.form</field>
-    <field name="model">stock.location</field>
-    <field name="inherit_id" ref="stock.view_location_form"/>
-    <field name="arch" type="xml">
-        <xpath expr="//field[@name='return_location']" position="after">
-            <field name="deposit_location"/>
-        </xpath>
-    </field>
-</record>
-
-<record id="deposit_quant_search_view" model="ir.ui.view">
-    <field name="name">deposit.stock.quant.search</field>
-    <field name="model">stock.quant</field>
-    <field name="inherit_id" ref="stock.quant_search_view"/>
-    <field name="arch" type="xml">
-        <xpath expr="//filter[@name='internal_loc']" position="after">
-            <!-- we need and operator -->
-            <separator/>
-            <filter name='deposit_loc'
-                    string="Deposit Locations"
-                    domain="[('location_id.deposit_location', '=', True)]"/>
-            <filter name='deposit_loc_exclude'
-                    string="Deposit Locations Excluded"
-                    domain="[('location_id.deposit_location', '!=', True)]"/>
-        </xpath>
-    </field>
-</record>
-
-<record id="deposit_product_open_quants" model="ir.actions.act_window">
-    <field name="name">Stock In Deposit</field>
-    <field name="context">{'search_default_internal_loc': 1,
-                           'search_default_locationgroup':1,
-                           'search_default_deposit_loc': 1}
-    </field>
-    <field name="domain">[('product_id', '=', active_id)]</field>
-    <field name="res_model">stock.quant</field>
-</record>
-
 <record id="deposit_view_move_search" model="ir.ui.view">
     <field name="name">deposit.stock.move.search</field>
     <field name="model">stock.move</field>
@@ -68,58 +29,6 @@
     <field name="name">Moves</field>
     <field name="res_model">stock.move</field>
     <field name="view_id" ref="stock.view_move_tree"/>
-</record>
-
-<record id="deposit_quantsact" model="ir.actions.act_window">
-    <field name="name">Deposit Quants</field>
-    <field name="res_model">stock.quant</field>
-    <field name="view_type">form</field>
-    <field name="view_mode">tree,form</field>
-    <field name="context">{'search_default_internal_loc': 1,
-                           'search_default_deposit_loc': 1,
-                           'group_by': ['product_id', 'location_id']}
-    </field>
-    <field name="help">
-        This analysis gives you a fast overview on the current stock level of
-        your products in deposit locations.
-    </field>
-</record>
-
-<menuitem id="menu_deposit_quants" name="Deposit Inventory Valuation"
-          parent="stock.menu_warehouse_report" sequence="125"
-          action="deposit_quantsact"/>
-
-<record id="deposit_view_picking_internal_search" model="ir.ui.view">
-    <field name="name">deposit.stock.picking.internal.search</field>
-    <field name="model">stock.picking</field>
-    <field name="inherit_id" ref="stock.view_picking_internal_search"/>
-    <field name="arch" type="xml">
-        <xpath expr="//field[@name='partner_id']" position="after">
-            <separator/>
-            <filter name='deposit_loc'
-                    string="Deposit Locations"
-                    domain="[('location_dest_id.deposit_location', '=', True)]"/>
-        </xpath>
-    </field>
-</record>
-
-<record id="deposit_view_picking_form" model="ir.ui.view">
-    <field name="name">deposit.stock.picking.form</field>
-    <field name="model">stock.picking</field>
-    <field name="inherit_id" ref="stock.view_picking_form"/>
-    <field name="arch" type="xml">
-        <xpath expr="//field[@name='picking_type_id']" position="after">
-            <field name="is_deposit" invisible="1"/>
-        </xpath>
-        <xpath expr="//field[@name='owner_id']" position="attributes">
-            <attribute name="attrs">
-                {'required': [
-                    ('picking_type_code', '=', 'internal'),
-                    ('is_deposit', '==', True),
-                ]}
-            </attribute>
-        </xpath>
-    </field>
 </record>
 
 </odoo>
